SQUASH???
(remove this part of the message while squashing)
Use ` `, not `[ ]`, as the proposed log message described.

(append to the proposed log message of the previous)
Also, do not write `\+` in BRE and expect it to mean 1 or more;
it is a GNU extension that may not work everywhere.

Remove '.*' at the end of a pattern that is not right-anchored.

Helped-by: Junio C Hamano <gitster@pobox.com>