csum-file: introduce discard_hashfile()
The hashfile API is used to write out a "hashfile", which has a
final checksum (typically SHA-1) at the end.  An in-core hashfile
structure has up to two file descriptors and a few buffers that can
only be freed by calling a helper function that is private to the
csum-file implementation.

The usual flow of a user of the API is to first obtain a hashfile by
calling either hashfd() or hashfd_check(), call hashwrite() number
of times, and then call finalize_hashfile() to conclude the access,
which writes the checksum, closes file descriptors and releases
associated buffers.  But there is no clean way for a caller that
detects an error and wants to abort the operation without calling
finalize_hashfile().  This resulted in "update-index" leaking the
hashfile structure it uses, for example, when it detects an error.

Introduce discard_hashfile() API function to allow them to release
the resources held by a hashfile structure the callers want to
dispose of, and use that in read-cache.c:do_write_index(), which is
a central place that writes the index file.

Mark t2107 as leak-free, as this leak in "update-index --cacheinfo"
that deliberately fails is now plugged.

Signed-off-by: Junio C Hamano <gitster@pobox.com>